requirements.txt
kmerator
snakemake
rdeer-service
PyYAML
pyfaidx
scipy
pandas
plotly
seaborn
